Update content display logics.
